requirements.txt
streamlit==1.52.2
pandas==2.3.3
ollama==0.6.1
joblib==1.5.3
langgraph==1.0.5
python-dotenv==1.2.1
faker==40.1.0
Pillow==12.1.0
reportlab==4.4.7
numpy==2.2.6
llama-index-core==0.14.12
SQLAlchemy==2.0.45
pydantic==2.12.5
pydantic_core==2.41.5
pypdf==6.5.0
matplotlib==3.10.8
seaborn==0.13.2
joblib==1.5.3
scikit-learn==1.8.0
xgboost==3.1.2
shap==0.50.0
fastapi==0.110.2
uvicorn==0.22.0
openpyxl==3.1.2
email-validator==2.1.1